+++ 00_Projects/00_Global/Libft/ft_memcpy.c
@@ -1,8 +1,8 @@
 #include <stddef.h>
 
-void *ft_memcpy(void *dest, const void *src, int n)
+void *ft_memcpy(void *dest, const void *src, size_t n)
 {
-    int i = 0;
+    size_t i = 0;
     unsigned char *d = (unsigned char *)dest;
     const unsigned char *s = (const unsigned char *)src;
