+++ 00_Projects/00_Global/Libft/ft_memset.c
@@ -1,6 +1,8 @@
-void *ft_memset(void *s, int c, int n)
+#include <stddef.h>
+
+void *ft_memset(void *s, int c, size_t n)
 {
-    int i = 0;
+    size_t i = 0;
     unsigned char *ptr = (unsigned char *)s;
 
     while (i < n)
